+++ server/src/main/java/com/continuuity/loom/scheduler/task/TaskQueueService.java
@@ -15,10 +15,13 @@
  */
 package com.continuuity.loom.scheduler.task;
 
+import com.continuuity.loom.admin.Tenant;
 import com.continuuity.loom.cluster.Node;
+import com.continuuity.loom.common.conf.Configuration;
 import com.continuuity.loom.common.conf.Constants;
 import com.continuuity.loom.common.queue.Element;
 import com.continuuity.loom.common.queue.QueueGroup;
+import com.continuuity.loom.common.queue.QueueMetrics;
 import com.continuuity.loom.common.queue.TrackingQueue;
 import com.continuuity.loom.http.request.FinishTaskRequest;
 import com.continuuity.loom.http.request.TakeTaskRequest;
@@ -26,6 +29,12 @@ import com.continuuity.loom.management.LoomStats;
 import com.continuuity.loom.provisioner.TenantProvisionerService;
 import com.continuuity.loom.store.cluster.ClusterStore;
 import com.continuuity.loom.store.cluster.ClusterStoreService;
+import com.continuuity.loom.store.tenant.TenantStore;
+import com.google.common.cache.CacheBuilder;
+import com.google.common.cache.CacheLoader;
+import com.google.common.cache.LoadingCache;
+import com.google.common.collect.Iterators;
+import com.google.common.collect.Maps;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.inject.Inject;
@@ -35,6 +44,7 @@ import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
 
 /**
  * Manages handing out tasks from task queue, and recording status after the task is done.
@@ -43,20 +53,24 @@ public class TaskQueueService {
   private static final Logger LOG = LoggerFactory.getLogger(TaskQueueService.class);
 
   private final ClusterStore clusterStore;
+  private final TenantStore tenantStore;
   private final TaskService taskService;
   private final NodeService nodeService;
   private final TenantProvisionerService tenantProvisionerService;
   private final LoomStats loomStats;
   private final QueueGroup taskQueues;
   private final QueueGroup jobQueues;
+  private final LoadingCache<String, QueueMetrics> queueMetricsCache;
 
   @Inject
-  private TaskQueueService(@Named(Constants.Queue.PROVISIONER) QueueGroup taskQueues,
+  private TaskQueueService(@Named(Constants.Queue.PROVISIONER) final QueueGroup taskQueues,
                            @Named(Constants.Queue.JOB) QueueGroup jobQueues,
                            ClusterStoreService clusterStoreService,
                            TenantProvisionerService tenantProvisionerService,
                            TaskService taskService,
                            NodeService nodeService,
+                           TenantStore tenantStore,
+                           Configuration conf,
                            LoomStats loomStats) {
     this.clusterStore = clusterStoreService.getSystemView();
     this.taskService = taskService;
@@ -65,6 +79,45 @@ public class TaskQueueService {
     this.loomStats = loomStats;
     this.taskQueues = taskQueues;
     this.jobQueues = jobQueues;
+    this.tenantStore = tenantStore;
+    final int queueCacheSeconds = conf.getInt(Constants.Metrics.QUEUE_CACHE_SECONDS);
+    // queue metrics can be expensive to fetch
+    this.queueMetricsCache = CacheBuilder.newBuilder()
+      .expireAfterWrite(queueCacheSeconds, TimeUnit.SECONDS)
+      .build(new CacheLoader<String, QueueMetrics>() {
+        @Override
+        public QueueMetrics load(String tenantId) {
+          int numQueued = Iterators.size(taskQueues.getQueued(tenantId));
+          int numInProgress = Iterators.size(taskQueues.getBeingConsumed(tenantId));
+          return new QueueMetrics(numQueued, numInProgress);
+        }
+      });
+  }
+
+  /**
+   * Get a snapshot of the number of queued and in progress elements for all queues in the task queues.
+   *
+   * @return Snapshot of the number of queued and in progress elements in the task queues
+   */
+  public Map<String, QueueMetrics> getTaskQueueMetricsSnapshot() throws IOException {
+    // get tenant list from the tenant store instead of from the queue group because queues are loaded lazily
+    // in the queue group and not derived from zookeeper, so we might not get all the queues.
+    Map<String, QueueMetrics> queueMetrics = Maps.newHashMap();
+    for (Tenant tenant : tenantStore.getAllTenants()) {
+      queueMetrics.put(tenant.getSpecification().getName(), queueMetricsCache.getUnchecked(tenant.getId()));
+    }
+    return queueMetrics;
+  }
+
+  /**
+   * Get the queue metrics for the given tenant.
+   *
+   * @param tenantId Id of the tenant to get queue metrics for
+   * @return Queue metrics for the tenants
+   * @throws IOException
+   */
+  public QueueMetrics getTaskQueueMetricsSnapshot(String tenantId) throws IOException {
+    return queueMetricsCache.getUnchecked(tenantId);
   }
 
   /**
@@ -78,7 +131,7 @@ public class TaskQueueService {
    * @throws IOException if there was an error persisting task information.
    */
   public String takeNextClusterTask(TakeTaskRequest takeRequest) throws IOException, MissingEntityException {
-    //loomStats.setQueueLength(taskQueues.size(queueName));
+    loomStats.setQueueLength(taskQueues.size());
     String queueName = takeRequest.getTenantId();
     String provisionerId = takeRequest.getProvisionerId();
     String workerId = takeRequest.getWorkerId();
@@ -136,8 +189,7 @@ public class TaskQueueService {
    * @throws IOException if there was an error persisting task information.
    */
   public void finishClusterTask(FinishTaskRequest finishRequest) throws MissingEntityException, IOException {
-    // TODO: implement per tenant queue statistics
-    //loomStats.setQueueLength(taskQueue.size());
+    loomStats.setQueueLength(taskQueues.size());
 
     String workerId = finishRequest.getWorkerId();
     String queueName = finishRequest.getTenantId();